document_verifier_service/requirements.txt
Flask>=2.0
pillow>=8.0
pytesseract>=0.3
opencv-python>=4.5
numpy>=1.19
itsdangerous>=2.0
pytest>=6.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
datasets>=2.7.0
scikit-learn>=1.2.0
pandas>=1.5.0
matplotlib>=3.6.0
tqdm>=4.64.0
onnx>=1.12.0
onnxruntime>=1.13.0
pdf2image>=1.16.0
